Creating the UI data handling
Given that I have the JSON data coming ing right now, I just need to put the data in my HTML template. In order to accomplish that, I created
a class that gets all the necessary HTML tags in the constructor and I also set a method that gets the JSON object as a parameter and
puts the data in their corresponding tags. It is worth mentioning that the image icon provided by the API needed to be appended to the
url src attribute.