Adding the Favicon and Navbar
In order to make a responsive Navbar, I used Bootstrap and added my own links and brand. The  "bg-transparent"
class allows me to have a navbar with a transparent background, so the users can see the bg image instead.

In the other hand, the favicon is in the root folder and the following HTML link tag allows me to use it in my
webpage:
<link rel="shortcut icon" type="image/x-icon" href="favicon.ico">